fix: a grep hit past 65535 chars into a line
A picker row's `(head, match start, match end, tag)` are char offsets into an
unbounded label, but every layer typed them as `u16` — the `btv._picker_push`
argument, `PickerPush.layout`, `RowLayout`, the wire parse, and a cast inside
`fit_row`. A live_grep hit past 65535 chars in (a minified bundle, a one-line
JSON blob) failed the conversion, which rejected the WHOLE batch: "error
converting Lua integer to u16 (out of range)", surfaced as an unhandled promise
rejection, and every row pushed with it was lost. Widened to `u32` end to end.

With the type fixed the search runs — and hangs. `rg` reports the full line
once per hit ON that line, so each row carried the whole megabyte into the item
list, across the bridge, and into every painted frame, for a body the widget
windows to the row's width anyway: half a second per keystroke on a 1MB line.
`live_grep` now clips an over-long body to a window around the hit (1024 bytes
either side, cut on a char boundary, each end marked `…`) and measures its char
range after clipping — `charlen` over a megabyte once per hit was the same
O(N^2). Same probe after: 66ms/100ms instead of 536ms/725ms. The cap is wider
than any terminal, so an ordinary source line is untouched.

Both tests fail with their fix reverted. The `u16` one pushes from a generic
source rather than live_grep, whose clip would otherwise keep the offsets small
enough to pass either way.

+++ crates/bemtvi-core/src/editor/menu.rs
@@ -169,19 +169,19 @@ pub type CmdlineCandidate = (String, String, Option<String>, Option<(usize, usiz
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub struct RowLayout {
     /// Char length of the leading location column (`"src/foo.rs:12:5: "`).
-    pub head: u16,
+    pub head: u32,
     /// The source's match within the label, as a half-open **char** range. Empty
     /// (`match_end == match_start`) when the source knows only where the interesting
     /// content begins — the body still windows around it, nothing is highlighted.
-    pub match_start: u16,
-    pub match_end: u16,
+    pub match_start: u32,
+    pub match_end: u32,
     /// Char length of a **pinned tag** at the head's start (`"E "` on a diagnostics
     /// row): the part that *classifies* the row, which survives the elision the rest
     /// of the head takes when the column is too narrow for it. Without it the head
     /// elides tail-first and the classification — the very thing the row is scanned
     /// by — would be the first thing to go. `0` for a head that is pure location
     /// (live_grep's `path:line:col: `), where the tail is what matters.
-    pub tag: u16,
+    pub tag: u32,
 }
 
 /// One candidate row: its display `label` and the **opaque source key** that
